+++ solutions/jetbrains.mps.samples.Physics.runtime/source_gen/jetbrains/mps/samples/Physics/java/runtime/objects/Function.java
@@ -4,5 +4,5 @@ package jetbrains.mps.samples.Physics.java.runtime.objects;
 
 
 public interface Function<T, S extends SystemScope> {
-  T compute(World world, S scope, PhysicalEntity currentEntity, long time);
+  T compute(World world, S scope, PhysicalEntity currentEntity, double time);
 }
